Move untaring into config
We move the untaring of the external files into the config module to be
done as the last step in the validation.
